Upping version number
The changes I have made to Graph and Vector2D are significant enough to
need a new version number on them.

Also so I have it somewhere. For this project the version number is
#.#.#
The first is the major version number and represents "releases"
The second is basically representing completed branches
The third is for significant completed changes within a branch
So 1.0.1 means we are after release 1 on the first branch and the first
significant change has been completed