New feature incoming
Was just reminded that for some fancy flee mechanics I need to be able
to rerun the map and not reset the values. Basically I need an option
that when processing the graph any pre-existing aggregate costs don't
get wiped. New tests first to make sure it ends up working.