This might fix it
So when about to go to sleep a thought hit me. I realized that my
problem is not in the graph processing but in my method of getting a
path. As it was there would be a prebaked answer for which edge provided
the shortest path so it was really quick and easy to just follow that
back to the source and return that path. With this double processing of
the map that no longer worked as the 'shortest' path didn't always exist
anymore as there were instances were the path needs to go 'uphill'. Thus
in a fit of "Not going to let sleep eat this idea" I coded a variation
on the getPathFrom method. How it works is as follows:

To make it a unique method I added a boolean input which can just let
you use the other method.

After that we make currentpoint and set it to the target then add it to
the list for the path

Now it goes into a while loop that continues until the current point is
a source point and thus we have a path.

First in the loop we get a list of all points connected to the current
point that are not null and not already in the path list.

Next is a quick error check were if the list of connected points is
empty we just break out of the loop.

Then we set a new point to the first point in the connected point list
then iterate through said list and if a point has a lower cost we set
newPoint to that one instead.

Finally for the loop we set currentPoint to newPoint and add it to the
path

And of course once the path reaches a source we return the path list.