needing to fix inherent problem with DijkstraGraphFactory
You can't get a point from the graph so I am switching the return value
to be the 2d array of points that gets made in it. To make the graph
itself the method now accepts a DijkstraGraph and through the magic of
references the one you provide will become the graph.