Boom, version 4 is go
I have changed a lot with this now. No optimization yet mind you but at
this point unless vitally needed I will avoid breaking backwards
compatability going forward at least for a while.

One of the big changes is that I went fully in on this being an
algorithm for the sort of tile layouts you see in a roguelike. This
comes mostly from only having the tiles themselves contain the cost to
enter them. This means the algorithm is assuming that no matter what
direction your coming from it will only be the next tile over.